src => lib sync
