+++ lib/utils.js
@@ -37,7 +37,7 @@ export function useAnJurors() {
           GQL_ENDPOINT,
           `
             {
-              jurors(first: 1000) {
+              jurors(first: 1000, where: { activeBalance_gt: 0 }) {
                 activeBalance
               }
             }
