+++ lib/web3-utils.js
@@ -167,6 +167,53 @@ export function formatUnits(
   return commas ? EthersUtils.commify(valueBeforeCommas) : valueBeforeCommas
 }
 
+export function useTokenReserve(tokenSymbol) {
+  const [tokenReserves, setTokenReserves] = useState(null)
+  const [loading, setLoading] = useState(false)
+  useEffect(() => {
+    let cancelled = false
+    let retryTimer
+    async function fetchTokenReserve() {
+      try {
+        const [tokenAddress] = getKnownContract(`TOKEN_${tokenSymbol}`)
+        if (!tokenAddress) {
+          throw new Error(`Unsupported token symbol: ${tokenSymbol}`)
+        }
+        setLoading(true)
+
+        const tokenReserves = await getTokenReserves(
+          tokenAddress,
+          Number(env('CHAIN_ID'))
+        )
+
+        if (!tokenReserves) {
+          throw new Error('Could not fetch reserves')
+        }
+
+        if (!cancelled) {
+          const reserve = bigNum(
+            tokenReserves.tokenReserve.amount.toFixed(0, 1).toString()
+          )
+          setTokenReserves(reserve)
+          setLoading(false)
+        }
+      } catch (e) {
+        setLoading(false)
+        retryTimer = setTimeout(fetchTokenReserve, UNISWAP_MARKET_RETRY_EVERY)
+      }
+
+      return () => {
+        cancelled = true
+        clearTimeout(retryTimer)
+      }
+    }
+
+    fetchTokenReserve()
+  }, [tokenSymbol])
+
+  return useMemo(() => [tokenReserves, loading], [loading, tokenReserves])
+}
+
 // Wraps uniswap’s getMarketDetails() to only require
 // the other token in the pair, qualified using its symbol.
 async function getAnjMarketDetails(tokenSymbol) {
@@ -181,12 +228,15 @@ async function getAnjMarketDetails(tokenSymbol) {
       if (!tokenAddress) {
         throw new Error(`Unsupported token symbol: ${symbol}`)
       }
-      const tokenData = getTokenReserves(tokenAddress, Number(env('CHAIN_ID')))
+      const tokenReserves = getTokenReserves(
+        tokenAddress,
+        Number(env('CHAIN_ID'))
+      )
 
-      if (!tokenData) {
+      if (!tokenReserves) {
         throw new Error('Could not fetch reserves')
       }
-      return tokenData
+      return tokenReserves
     })
   )
